增加地图名过滤+断连提醒

+++ app/src/main/java/ddr/example/com/nddrandroidclient/entity/info/MapFileStatus.java
@@ -98,7 +98,7 @@ public class MapFileStatus {
                 targetPoint.setName(targetPtItems.get(i).getPtName().toStringUtf8());
                 targetPoint.setX(targetPtItems.get(i).getPtData().getX());
                 targetPoint.setY(targetPtItems.get(i).getPtData().getY());
-                targetPoint.setTheta(targetPtItems.get(i).getPtData().getTheta());
+                targetPoint.setTheta((int) targetPtItems.get(i).getPtData().getTheta());
                 cTargetPoints.add(targetPoint);
             }
             for (int i = 0; i < pathLineItemExes.size(); i++) {
@@ -172,7 +172,7 @@ public class MapFileStatus {
             targetPoint.setName(targetPtItems.get(i).getPtName().toStringUtf8());
             targetPoint.setX(targetPtItems.get(i).getPtData().getX());
             targetPoint.setY(targetPtItems.get(i).getPtData().getY());
-            targetPoint.setTheta(targetPtItems.get(i).getPtData().getTheta());
+            targetPoint.setTheta((int)targetPtItems.get(i).getPtData().getTheta());
             targetPoints.add(targetPoint);
         }
         for (int i = 0; i < pathLineItemExes.size(); i++) {
